fix: update public battle view
