+++ apps/web/src/features/public-battle/public-battle.tsx
@@ -1,17 +1,17 @@
+import { RouteErrorState } from "@/components/router/route-error-state";
+import { PublicBattleSkeleton } from "@/features/public-battle/public-battle-skeleton";
+import { BattleDetailView } from "@/features/user/battle-panel/battle-panel-single-battle/components/battle-detail-view";
 import {
   usePublicBattlesControllerGetPublicBattle,
   usePublicBattlesControllerGetPublicBattleRaw,
+  usePublicBattlesControllerGetPublicBattleTimeline,
 } from "@/lib/api/generated/battlelog/public-battles/public-battles";
-import { useParams } from "@tanstack/react-router";
-import { ScrollArea } from "@lootlog/ui/components/scroll-area";
 import {
-  BattleLog,
-  BattleOverviewCard,
-  BattleStatsTable,
-} from "@/components/battle";
-import { PublicBattleSkeleton } from "./public-battle-skeleton";
-
-const CDN_BASE_URL = "https://micc.garmory-cdn.cloud/obrazki/postacie";
+  getRouteErrorMessage,
+  getRouteErrorStatus,
+  normalizeRouteErrorStatus,
+} from "@/lib/router/route-errors";
+import { useParams } from "@tanstack/react-router";
 
 export const PublicBattle = () => {
   const { id: battleId } = useParams({ from: "/battles/$id" });
@@ -21,42 +21,40 @@ export const PublicBattle = () => {
     isLoading: isBattleLoading,
     error: battleError,
   } = usePublicBattlesControllerGetPublicBattle({ battleId });
-  const { data: rawBattle, isLoading: isRawBattleLoading } =
-    usePublicBattlesControllerGetPublicBattleRaw({ battleId });
+  const {
+    data: rawBattle,
+    isLoading: isRawBattleLoading,
+    error: rawBattleError,
+  } = usePublicBattlesControllerGetPublicBattleRaw({ battleId });
+  const { data: timeline, isPending: isTimelinePending } =
+    usePublicBattlesControllerGetPublicBattleTimeline({ battleId });
 
-  const isLoading = isBattleLoading || isRawBattleLoading;
-
-  if (isLoading) {
+  if (isBattleLoading || isRawBattleLoading) {
     return <PublicBattleSkeleton />;
   }
 
-  if (battleError || !battle || !rawBattle) {
+  const error = battleError ?? rawBattleError;
+
+  if (error || !battle || !rawBattle) {
     return (
-      <div className="min-h-screen flex items-center justify-center bg-background">
-        <div className="text-center space-y-4">
-          <h1 className="text-4xl font-bold text-foreground">
-            Nie znaleziono walki
-          </h1>
-          <p className="text-muted-foreground">
-            {battleError?.message ||
-              "Przepraszamy, nie mogliśmy znaleźć tej walki."}
-          </p>
-        </div>
+      <div className="flex min-h-dvh bg-background">
+        <RouteErrorState
+          status={normalizeRouteErrorStatus(getRouteErrorStatus(error))}
+          description={getRouteErrorMessage(error)}
+        />
       </div>
     );
   }
 
   return (
-    <div className="h-dvh bg-background">
-      <ScrollArea className="h-full">
-        <BattleOverviewCard
-          battle={battle}
-          cdnBaseUrl={CDN_BASE_URL}
-          showActions={false}
-        />
-        <BattleStatsTable battle={battle} className="sm:max-w-full" />
-        <BattleLog rawBattle={rawBattle.rawData} warriors={battle.warriors} />
-      </ScrollArea>
+    <div className="h-dvh min-h-0 overflow-hidden bg-background">
+      <BattleDetailView
+        battle={battle}
+        battleId={battleId}
+        isTimelinePending={isTimelinePending}
+        rawBattle={rawBattle.rawData}
+        timeline={timeline}
+      />
     </div>
   );
 };
